+++ src/main/java/gs/mclo/api/internal/LimitedReader.java
@@ -1,5 +1,6 @@
 package gs.mclo.api.internal;
 
+import gs.mclo.api.response.Limits;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
@@ -9,6 +10,7 @@ import java.util.Objects;
 /**
  * A Helper class that limits the number of bytes and/or lines read from a Reader.
  */
+@SuppressWarnings("SynchronizeOnNonFinalField")
 public final class LimitedReader extends Reader {
     /**
      * The underlying reader
@@ -18,29 +20,27 @@ public final class LimitedReader extends Reader {
     /**
      * The remaining byte limit. If null, no limit is enforced.
      */
-    @Nullable
     private Integer remainingByteLimit;
     /**
      * The remaining line limit. If null, no limit is enforced.
      */
-    @Nullable
     private Integer remainingLineLimit;
 
     /**
      * Creates a new LimitedReader with the given limits.
-     * @param in the underlying reader
-     * @param byteLimit the maximum number of bytes to read. If null, no limit is enforced.
-     * @param lineLimit the maximum number of lines to read. If null, no limit is enforced.
+     *
+     * @param in     the underlying reader
+     * @param limits the limits to enforce. If null, no limits are enforced.
      */
-    public LimitedReader(Reader in, @Nullable Integer byteLimit, @Nullable Integer lineLimit) {
+    public LimitedReader(Reader in, Limits limits) {
         this.in = Objects.requireNonNull(in);
-        this.remainingByteLimit = byteLimit;
-        this.remainingLineLimit = lineLimit;
+        this.remainingByteLimit = limits.getMaxLength();
+        this.remainingLineLimit = limits.getMaxLines();
     }
 
     @Override
     public int read(char[] chars, int offset, int length) throws IOException {
-        synchronized(this.lock) {
+        synchronized (this.lock) {
             if (this.in == null) {
                 throw new IOException("Stream closed");
             }
@@ -53,21 +53,19 @@ public final class LimitedReader extends Reader {
 
             int i = 0;
             for (; i < read; i++) {
-                if (this.remainingByteLimit != null && this.remainingByteLimit <= 0) {
+                if (this.remainingByteLimit <= 0) {
                     return i == 0 ? -1 : i;
                 }
 
                 char c = readChars[i];
                 var remainingByteLimit = this.remainingByteLimit;
-                if (remainingByteLimit != null) {
-                    int byteCount = getPartialUTF8Size(c);
-                    if (this.getRemainingUTF8Size(c) > remainingByteLimit) {
-                        return i == 0 ? -1 : i;
-                    }
-                    this.remainingByteLimit -= byteCount;
+                int byteCount = getPartialUTF8Size(c);
+                if (this.getRemainingUTF8Size(c) > remainingByteLimit) {
+                    return i == 0 ? -1 : i;
                 }
+                this.remainingByteLimit -= byteCount;
 
-                if (this.remainingLineLimit != null && c == '\n') {
+                if (c == '\n') {
                     this.remainingLineLimit--;
                     if (this.remainingLineLimit <= 0) {
                         return i == 0 ? -1 : i;
@@ -83,7 +81,7 @@ public final class LimitedReader extends Reader {
 
     @Override
     public void close() throws IOException {
-        synchronized(this.lock) {
+        synchronized (this.lock) {
             if (this.in != null) {
                 try {
                     this.in.close();
@@ -97,6 +95,7 @@ public final class LimitedReader extends Reader {
     /**
      * Get the number of bytes it would require to encode this full char in UTF-8.
      * High surrogates are considered 4 bytes since the low surrogate is still needed.
+     *
      * @param character the UTF-16 input character
      * @return size required to encode this character in UTF-8
      */
@@ -115,6 +114,7 @@ public final class LimitedReader extends Reader {
     /**
      * Get the number of bytes it would require to encode this (part of a) char in UTF-8.
      * Surrogates are considered to be 2 bytes by this method.
+     *
      * @param character the UTF-16 input character
      * @return size this char would take in UTF-8
      */